feat: better linking logic for C++ stdlib
