+++ build.rs
@@ -11,7 +11,16 @@ fn main() {
             .define("CHARLS_BUILD_SAMPLES", "0")
             .always_configure(true)
             .build();
+
         println!("cargo:rustc-link-search=native={}/lib", dst.display());
+        println!(
+            "cargo:rustc-link-lib={}=charls",
+            if cfg!(feature = "static") {
+                "static"
+            } else {
+                "dylib"
+            }
+        );
     }
 
     #[cfg(feature = "vcpkg")]
@@ -19,26 +28,4 @@ fn main() {
         .emit_includes(true)
         .find_package("charls")
         .unwrap();
-
-    #[cfg(feature = "static")]
-    {
-        println!("cargo:rustc-link-lib=static=charls");
-    }
-
-    #[cfg(not(feature = "static"))]
-    {
-        println!("cargo:rustc-link-lib=charls");
-    }
-
-    if let Ok(inner) = std::env::var("CARGO_CFG_TARGET_OS") {
-        match inner.as_str() {
-            "linux" => {
-                println!("cargo:rustc-link-lib=stdc++");
-            }
-            "macos" => {
-                println!("cargo:rustc-link-lib=c++");
-            }
-            _ => {}
-        }
-    }
 }
